+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteContent/PaletteContentText.cs
@@ -99,7 +99,7 @@ namespace Krypton.Toolkit
         /// Initialize a new instance of the PaletteContentText class.
         /// </summary>
         /// <param name="needPaint">Delegate for notifying paint requests.</param>
-        public PaletteContentText(NeedPaintHandler needPaint) =>
+        public PaletteContentText(NeedPaintHandler? needPaint) =>
             // Store the provided paint notification delegate
             NeedPaint = needPaint;